requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
peewee==3.10.0
pycparser==2.19
python-decouple==3.1
python-telegram-bot==12.0.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
tornado==6.0.3
urllib3==1.25.3
